refactoring of the rl code
